+++ src/lib.rs
@@ -348,6 +348,7 @@
 //! * `toml`: enables TOML support (`assert_toml_snapshot!`)
 //! * `redactions`: enables support for redactions
 //! * `glob`: enables support for globbing (`glob!`)
+//! * `colors`: enables color output (enabled by default)
 //!
 //! # Settings
 //!